Create AccessibleText type,  common form props, form text components. Create Forms folder and restructure
