+++ packages/components/src/components/Forms/Checkbox/Checkbox.stories.tsx
@@ -3,7 +3,7 @@ import { View } from 'react-native'
 import React, { useState } from 'react'
 
 import { Checkbox, CheckboxProps } from './Checkbox'
-import { generateDocs } from '../../utils/storybook'
+import { generateDocs } from '../../../utils/storybook'
 
 const meta: Meta<CheckboxProps> = {
   title: 'Checkbox',
@@ -43,16 +43,11 @@ type Story = StoryObj<CheckboxProps>
 const statefulComponentRenderer = (props: CheckboxProps) => {
   const {
     description,
-    descriptionA11y,
     error,
-    errorA11y,
     header,
-    headerA11y,
     hint,
-    hintA11y,
     indeterminate,
     label,
-    labelA11y,
     required,
     tile,
   } = props
@@ -63,16 +58,11 @@ const statefulComponentRenderer = (props: CheckboxProps) => {
     <Checkbox
       checked={checked}
       description={description}
-      descriptionA11y={descriptionA11y}
       error={error}
-      errorA11y={errorA11y}
       header={header}
-      headerA11y={headerA11y}
       hint={hint}
-      hintA11y={hintA11y}
       indeterminate={indeterminate}
       label={label}
-      labelA11y={labelA11y}
       onPress={() => setChecked(!checked)}
       required={required}
       tile={tile}
@@ -80,28 +70,25 @@ const statefulComponentRenderer = (props: CheckboxProps) => {
   )
 }
 
-const header = 'Label Header',
-  headerA11y = 'Accessibility override for header',
-  hint = 'Hint text',
-  hintA11y = 'Accessibility override for hint',
-  error = 'Error text',
-  errorA11y = 'Accessibility override for error',
-  label = 'Label',
-  labelA11y = 'Accessibility override for label',
-  description = 'Checkbox description',
-  descriptionA11y = 'Accessibility override for description'
+const header = {
+    text: 'Label Header',
+    a11y: 'Accessibility override for header',
+  },
+  hint = { text: 'Hint text', a11y: 'Accessibility override for hint' },
+  error = { text: 'Error text', a11y: 'Accessibility override for error' },
+  label = { text: 'Label', a11y: 'Accessibility override for label' },
+  description = {
+    text: 'Checkbox description',
+    a11y: 'Accessibility override for description',
+  }
 
 export const _Default: Story = {
   render: statefulComponentRenderer,
   args: {
     header,
-    headerA11y,
     hint,
-    hintA11y,
     label,
-    labelA11y,
     description,
-    descriptionA11y,
     required: true,
   },
 }
@@ -111,13 +98,9 @@ export const __Tile: Story = {
   args: {
     tile: true,
     header,
-    headerA11y,
     hint,
-    hintA11y,
     label,
-    labelA11y,
     description,
-    descriptionA11y,
   },
 }
 
@@ -125,7 +108,6 @@ export const ___CheckboxOnly: Story = {
   render: statefulComponentRenderer,
   args: {
     label,
-    labelA11y,
   },
 }
 
@@ -133,14 +115,9 @@ export const ____Error: Story = {
   render: statefulComponentRenderer,
   args: {
     header,
-    headerA11y,
     hint,
-    hintA11y,
     error,
-    errorA11y,
     label,
-    labelA11y,
     description,
-    descriptionA11y,
   },
 }
