+++ packages/components/src/components/Forms/CheckboxGroup/CheckboxGroup.tsx
@@ -1,22 +1,21 @@
-import { Text, View, ViewStyle } from 'react-native'
+import { View, ViewStyle } from 'react-native'
 import { spacing } from '@department-of-veterans-affairs/mobile-tokens'
-import { useTranslation } from 'react-i18next'
 import React, { FC, Fragment } from 'react'
 
+import { AccessibleText } from '../../../types/common'
 import { Checkbox } from '../Checkbox/Checkbox'
-import { ComponentWrapper } from '../../wrapper'
-import { Spacer } from '../Spacer/Spacer'
-import { useTheme } from '../../utils'
+import { ComponentWrapper } from '../../../wrapper'
+import { Error, Header, Hint } from '../FormText'
+import { Spacer } from '../../Spacer/Spacer'
+import { useTheme } from '../../../utils'
 
-export type CheckboxGroupItem = {
-  label: string
-  a11y?: string
+export type CheckboxGroupItem = AccessibleText & {
   value?: string | number
 }
 
 export type CheckboxGroupProps = {
   /** Array of checkbox options. Can be an array of strings or objects if values and/or a11y overrides are needed */
-  items: CheckboxGroupItem[] | string[]
+  items: CheckboxGroupItem[]
   /** OnPress logic to alter `checked` state or other behavior associated with the checkbox */
   onSelectionChange: (selected: (string | number)[]) => void
   /** Callback function that receives an updated array of selected values when checkboxes are pressed */
@@ -47,16 +46,12 @@ export const CheckboxGroup: FC<CheckboxGroupProps> = ({
   items,
   selectedItems,
   error,
-  errorA11y,
   header,
-  headerA11y,
   hint,
-  hintA11y,
   onSelectionChange,
   required,
   tile,
 }) => {
-  const { t } = useTranslation()
   const theme = useTheme()
 
   const handleCheckboxChange = (value: string | number) => {
@@ -67,34 +62,6 @@ export const CheckboxGroup: FC<CheckboxGroupProps> = ({
     } else {
       onSelectionChange([...selectedItems, value])
     }
-  }
-
-  /**
-   * Fonts
-   */
-  // TODO: Replace with typography tokens
-  const fontRegular = 'SourceSansPro-Regular'
-  const fontBold = 'SourceSansPro-Bold'
-
-  const fontHeader = {
-    fontFamily: fontRegular,
-    color: theme.vadsColorForegroundDefault,
-    fontSize: 20,
-    lineHeight: 30,
-  }
-
-  const fontHint = {
-    fontFamily: fontRegular,
-    color: theme.vadsColorForegroundSubtle,
-    fontSize: 16,
-    lineHeight: 22,
-  }
-
-  const fontError = {
-    fontFamily: fontBold,
-    color: theme.vadsColorForegroundError,
-    fontSize: 16,
-    lineHeight: 22,
   }
 
   /**
@@ -113,76 +80,37 @@ export const CheckboxGroup: FC<CheckboxGroupProps> = ({
     }
   }
 
-  /**
-   * Header
-   */
-  const headerProps = {
-    'aria-label': headerA11y,
-    style: fontHeader,
-  }
-
-  const requiredStyle = {
-    color: theme.vadsColorForegroundError,
-  }
-
-  const _header = header && (
-    <>
-      <Text {...headerProps}>
-        {header}
-        {required && (
-          <Text style={requiredStyle}>{` (*${t('required')})`}</Text>
-        )}
-      </Text>
-      <Spacer size="xs" />
-    </>
-  )
-
-  /**
-   * Hint
-   */
-  const hintProps = {
-    'aria-label': hintA11y,
-    style: fontHint,
-  }
-
-  const _hint = hint && (
-    <>
-      <Text {...hintProps}>{hint}</Text>
-      <Spacer size="xs" />
-    </>
-  )
-
-  /**
-   * Error
-   */
-  const errorProps = {
-    'aria-label': `${t('error')}: ${errorA11y || error}`,
-    style: fontError,
-  }
-
-  const _error = error && (
-    <>
-      <Text {...errorProps}>{error}</Text>
-      <Spacer size="xs" />
-    </>
-  )
-
   return (
     <ComponentWrapper>
       <View style={containerStyle}>
-        {_header}
-        {_hint}
-        {_error}
+        {header && (
+          <>
+            <Header text={header} required={required} />
+            <Spacer size="xs" />
+          </>
+        )}
+
+        {hint && (
+          <>
+            <Hint text={hint} />
+            <Spacer size="xs" />
+          </>
+        )}
+
+        {error && (
+          <>
+            <Error text={error} />
+            <Spacer size="xs" />
+          </>
+        )}
         {items.map((item, index) => {
-          const isObject = typeof item === 'object'
-          const label = isObject ? item.label : item
-          const value = isObject ? item.value || item.label : item
+          const value =
+            typeof item === 'object' ? item.value || item.text : item
 
           return (
             <Fragment key={`checkbox-group-item-${index}`}>
               <Checkbox
-                label={label}
-                labelA11y={isObject ? item.a11y : undefined}
+                label={item}
                 checked={selectedItems.includes(value)}
                 onPress={() => handleCheckboxChange(value)}
                 tile={tile}
