+++ packages/components/src/index.tsx
@@ -9,7 +9,7 @@ if (expoApp && App.initiateExpo) {
 // Export components here so they are exported through npm
 export { Alert } from './components/Alert/Alert'
 export { Button, ButtonVariants } from './components/Button/Button'
-export { Checkbox } from './components/Checkbox/Checkbox'
+export { Checkbox } from './components/Forms/Checkbox/Checkbox'
 export { Icon } from './components/Icon/Icon'
 export { Link } from './components/Link/Link'
 export { LoadingIndicator } from './components/LoadingIndicator/LoadingIndicator'
