+++ packages/components/src/types/common.ts
@@ -1,0 +1,8 @@
+export type AccessibleText =
+  | string
+  | {
+      /** Text to display */
+      text: string
+      /** Accessibility override for text */
+      a11y?: string
+    }
